# api-contract.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gourmet Guide AI API",
    "description": "API for the Gourmet Guide AI food recommendation service",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@gourmetguide.ai"
    }
  },
  "servers": [
    {
      "url": "https://api.gourmetguide.ai/v1",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.gourmetguide.ai/v1",
      "description": "Staging server"
    }
  ],
...
  "paths": {
    "/location/geocode": {
      "post": {
        "tags": ["location"],
        "summary": "Convert address to coordinates",
        "description": "Convert a text address to geographic coordinates",
        "operationId": "geocodeAddress",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoordinatesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid address supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/location/reverse-geocode": {
      "post": {
        "tags": ["location"],
        "summary": "Convert coordinates to address",
        "description": "Convert geographic coordinates to a text address",
        "operationId": "reverseGeocodeCoordinates",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoordinatesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid coordinates supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddressRequest": {
        "type": "object",
        "required": ["address"],
        "properties": {
          "address": {
            "type": "string",
            "example": "Jakarta, Indonesia"
          }
        }
      },
      "CoordinatesResponse": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double",
            "example": -6.2088
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "example": 106.8456
          },
          "formattedAddress": {
            "type": "string",
            "example": "Jakarta, Indonesia"
          }
        }
      },
      "CoordinatesRequest": {
        "type": "object",
        "required": ["latitude", "longitude"],
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double",
            "example": -6.2088
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "example": 106.8456
          }
        }
      },
      "AddressResponse": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "Jakarta, Indonesia"
          },
          "city": {
            "type": "string",
            "example": "Jakarta"
          },
          "country": {
            "type": "string",
            "example": "Indonesia"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "The requested resource was not found."
              },
              "details": {
                "type": "string",
                "example": "User not found."
              },
              "code": {
                "type": "string",
                "example": "NF_01"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}